// src/app/models/directory.model.ts
export class directory {
    id?: string;
    userId?: string;
    directoryCategoryId?: string;
    directoryCategoryDescription?: string;
    companyName?: string;
    description?: string;
    email?: string;
    phone?: string;
    website?: string;
    instagram?: string;
    facebook?: string;
    twitter?: string;
    addressLat?: number;
    addressLon?: number;
    addressLabel?: string;
    avatar?: string;
    statusId?: string;
    statusDescription?: string;
    count?: number;
}